Removed EventsBaseService as it is no longer user
